refactor: rename uuid58.ts to generate.ts
