+++ benchmark.ts
@@ -1,7 +1,7 @@
 import { UUID58_ALPHABET, UUID58_LENGTH } from "./alphabet.ts";
 import { uuid58Decode } from "./decode.ts";
 import { uuid58Encode } from "./encode.ts";
-import { uuid58 } from "./uuid58.ts";
+import { uuid58 } from "./generate.ts";
 
 Deno.bench("uuid58Encode", () => {
   uuid58Encode("f4b247fd-1f87-45d4-aa06-1c6fc0a8dfaf");
