+++ index.ts
@@ -1,4 +1,4 @@
 export { UUID58_ALPHABET, UUID58_REGEX } from "./alphabet.ts";
 export * from "./decode.ts";
 export * from "./encode.ts";
-export * from "./uuid58.ts";
+export * from "./generate.ts";
